+++ WebApplication/src/components/InstanceListExtended.vue
@@ -2,10 +2,11 @@
 import InstanceItem from "./InstanceItem.vue"
 import api from "../orthancApi"
 import { ObserveVisibility as vObserveVisibility } from 'vue3-observe-visibility'
+import SelectionStatus from "../helpers/selection-status";
 
 
 export default {
-    props: ['seriesId', 'seriesMainDicomTags', 'studyMainDicomTags', 'patientMainDicomTags', 'seriesInstances'],
+    props: ['seriesId', 'seriesMainDicomTags', 'studyMainDicomTags', 'patientMainDicomTags', 'seriesInstances', 'studyId', 'studySeries'],
     emits: ['deletedInstance'],
     data() {
         return {
@@ -22,6 +23,20 @@ export default {
             } else {
                 return [];
             }
+        },
+        selectedInstancesCount() {
+            const count = this.$store.getters['selection/getSelectedInstancesCount'](this.studyId, this.seriesId, this.seriesInstances);
+            if (count == 0) {
+                return "";
+            } else {
+                return count;
+            }
+        },
+        allSelected() {
+            return this.$store.getters['selection/getSeriesSelectionStatus'](this.studyId, this.seriesId) != SelectionStatus.NOT_SELECTED;
+        },
+        isPartialySelected() {
+            return this.$store.getters['selection/getSeriesSelectionStatus'](this.studyId, this.seriesId) == SelectionStatus.PARTIAL;
         },
     },
     watch: {
@@ -54,7 +69,14 @@ export default {
                 }
             }
 
-        }
+        },
+        clickSelectAll() {
+            if (this.allSelected == '' || !this.allSelected) { // this is the value before the click
+                this.$store.dispatch('selection/selectSeries', { studyId: this.studyId, seriesId: this.seriesId, studySeries: this.studySeries, isSelected: true });
+            } else {
+                this.$store.dispatch('selection/selectSeries', { studyId: this.studyId, seriesId: this.seriesId, studySeries: this.studySeries, isSelected: false });
+            }
+        },
     },
     components: { InstanceItem }
 }
@@ -64,7 +86,14 @@ export default {
     <table class="table table-responsive table-sm instance-table">
         <thead>
             <tr>
-                <th width="2%" scope="col" class="instance-table-header"></th>
+                <th width="2%" scope="col" class="instance-table-header">
+                    <div class="form-check" style="margin-left: 0.5rem">
+                        <input class="form-check-input" type="checkbox" :checked="allSelected"
+                            :indeterminate="isPartialySelected" @click="clickSelectAll"><span
+                            style="font-weight: 400; font-size: small;">{{ selectedInstancesCount
+                            }}</span>
+                    </div>
+                </th>
                 <th width="7%" scope="col" class="instance-table-header cut-text" data-bs-toggle="tooltip"
                     :title="$t('dicom_tags.InstanceNumber')">{{ $t('dicom_tags.InstanceNumber') }}</th>
                 <th width="40%" scope="col" class="instance-table-header cut-text" data-bs-toggle="tooltip"
@@ -77,6 +106,8 @@ export default {
             :instanceId="instanceId" :instanceInfo="instancesInfo[instanceId]"
             :studyMainDicomTags="this.studyMainDicomTags" :seriesMainDicomTags="this.seriesMainDicomTags"
             :patientMainDicomTags="this.patientMainDicomTags" @deletedInstance="onDeletedInstance"
+            :studyId="this.studyId" :studySeries="this.studySeries" :seriesInstances="this.seriesInstances"
+            :seriesId="this.seriesId"
             v-observe-visibility="{ callback: visibilityChanged, once: true }">
         </InstanceItem>
     </table>